chore: extract util functions into separate file
