+++ app/component/itinerary/navigator/hooks/useRealtimeLegs.js
@@ -6,149 +6,15 @@ import { fetchQuery } from 'react-relay';
 import { updateLatestNavigatorItineraryParams } from '../../../../store/localStorage';
 import { GeodeticToEcef, GeodeticToEnu } from '../../../../util/geo-utils';
 import { legTime } from '../../../../util/legUtils';
-import { epochToIso } from '../../../../util/timeUtils';
 import { legQuery } from '../../queries/LegQuery';
 import { getRemainingTraversal, validateTransitLeg } from '../NaviUtils';
-
-function nextTransitIndex(legs, i) {
-  for (let j = i; j < legs.length; j++) {
-    if (legs[j].transitLeg) {
-      return j;
-    }
-  }
-  // negative indicates not found
-  return -1;
-}
-
-function getLegGap(legs, index) {
-  return legTime(legs[index + 1].start) - legTime(legs[index].end);
-}
-
-// change non-transit legs' scheduled times
-function shiftLegs(legs, i1, i2, gap) {
-  for (let j = i1; j <= i2; j++) {
-    const leg = legs[j];
-    if (!leg.freezeStart) {
-      leg.start.scheduledTime = epochToIso(legTime(leg.start) + gap);
-    }
-    if (!leg.freezeEnd) {
-      leg.end.scheduledTime = epochToIso(legTime(leg.end) + gap);
-    }
-  }
-}
-
-// shift transit leg (for debugging)
-function shiftLeg(leg, gap) {
-  if (!leg.freezeStart) {
-    leg.start = {
-      ...leg.start,
-      estimated: {
-        ...leg.start.estimated,
-        time: epochToIso(legTime(leg.start) + gap),
-        delay: gap,
-      },
-    };
-  }
-  if (!leg.freezeEnd) {
-    leg.end = {
-      ...leg.end,
-      estimated: {
-        ...leg.end.estimated,
-        time: epochToIso(legTime(leg.end) + gap),
-        delay: gap,
-      },
-    };
-    leg.realtimeState = 'UPDATED';
-  }
-}
-
-// scale non-transit legs' scheduled times
-function scaleLegs(legs, i1, i2, k) {
-  const base = legTime(legs[i1].start);
-  for (let j = i1; j <= i2; j++) {
-    const leg = legs[j];
-    const s = legTime(leg.start);
-    const e = legTime(leg.end);
-    if (!leg.freezeStart) {
-      leg.start.scheduledTime = epochToIso(base + k * (s - base));
-    }
-    if (!leg.freezeEnd) {
-      leg.end.scheduledTime = epochToIso(base + k * (e - base));
-    }
-  }
-}
-
-function matchLegEnds(legs) {
-  if (legs.length < 2) {
-    return;
-  }
-  let transit;
-  let gap;
-
-  // shift first legs to match transit start
-  transit = nextTransitIndex(legs, 0);
-  if (transit > 0) {
-    gap = getLegGap(legs, transit - 1);
-    if (gap) {
-      shiftLegs(legs, 0, transit - 1, gap);
-    }
-  }
-
-  // shift transfers and legs after transit end
-  while (transit > 0) {
-    const walk = transit + 1; // first leg after transit
-    const nextTransit = nextTransitIndex(legs, walk);
-    const shiftEnd = nextTransit > 0 ? nextTransit - 1 : legs.length - 1;
-    if (shiftEnd > transit) {
-      gap = getLegGap(legs, transit);
-      if (gap) {
-        shiftLegs(legs, walk, shiftEnd, -gap);
-      }
-    }
-    if (nextTransit > walk) {
-      // check if transfer needs scaling
-      gap = getLegGap(legs, nextTransit - 1);
-      if (gap < 0) {
-        // transfer overlaps next transit leg, so we must make it shorter
-        const transferDuration =
-          legTime(legs[shiftEnd].end) - legTime(legs[walk].start);
-        scaleLegs(
-          legs,
-          walk,
-          shiftEnd,
-          (transferDuration + gap) / transferDuration,
-        );
-      }
-    }
-    transit = nextTransit;
-  }
-}
-
-function getLegsOfInterest(legs, now) {
-  if (!legs?.length) {
-    return {
-      firstLeg: undefined,
-      lastLeg: undefined,
-      currentLeg: undefined,
-      nextLeg: undefined,
-    };
-  }
-  const currentLeg = legs.find(
-    ({ start, end }) => legTime(start) <= now && legTime(end) >= now,
-  );
-
-  const nextStart = currentLeg ? legTime(currentLeg.end) : now;
-  const nextLeg = legs.find(({ start }) => legTime(start) >= nextStart);
-  const previousLeg = legs.findLast(({ end }) => legTime(end) < now);
-
-  return {
-    firstLeg: legs[0],
-    lastLeg: legs[legs.length - 1],
-    previousLeg,
-    currentLeg,
-    nextLeg,
-  };
-}
+import {
+  getLegsOfInterest,
+  matchLegEnds,
+  nextTransitIndex,
+  shiftLeg,
+  shiftLegs,
+} from './utils/realtimeLegUtils';
 
 function getInitialState(legs) {
   const time = Date.now();
